FIXED CLI INTEGRATION
