+++ src/cli/__init__.py
@@ -1,0 +1,2 @@
+"""CLI module for command-line interface components."""
+
